Create Call log Analysis
Provided with a call log history represented as a list of dictionaries. Each dictionary contains the following keys:

name (str): The name of the caller.
duration (int): The duration of the call in seconds.
There will be multiple entries with the same name in the dataset. Implement a function process_call_logs(call_logs, task) where task is one of the following: